+++ CHANGELOG.md
@@ -39,6 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added 'other' checkbox to 'Create Catalogue by importing a file' (for selecting custom piplelines)
 - Command SetExtractionIdentifier now supports changing the linkage identifier for specific ExtractionConfigurations only
 - Added new command `AlterTableMakeDistinct`
+- Added CLI GUI window for running Pipelines that displays progress
 
 ### Fixed